requirements.txt
smartapi-python==1.0.6
pyotp==2.8.0
requests==2.31.0
python-telegram-bot==20.0a4
Flask==2.3.2
gunicorn==21.2.0
